use uint64_t for nodes

+++ game.h
@@ -33,7 +33,7 @@ typedef struct {
     // Per engine, by index in engines[] array (not the same as color)
     int64_t movetime[2], time[2], increment[2];
     int movestogo[2];
-    unsigned nodes[2];
+    uint64_t nodes[2];
     int depth[2];
 
     int resignCount, resignScore;
