+++ options.c
@@ -82,8 +82,8 @@ Options options_new(int argc, const char **argv)
             else if (!strcmp(argv[i - 1], "-nodes")) {
                 str_t nodes[2] = {str_new(), str_new()};
                 split_engine_option(argv[i], nodes);
-                o.go.nodes[0] = atoi(nodes[0].buf);
-                o.go.nodes[1] = atoi(nodes[1].buf);
+                o.go.nodes[0] = atoll(nodes[0].buf);
+                o.go.nodes[1] = atoll(nodes[1].buf);
                 str_delete(&nodes[0], &nodes[1]);
             } else if (!strcmp(argv[i - 1], "-depth")) {
                 str_t depth[2] = {str_new(), str_new()};
